# app/openapi.yaml
openapi: 3.0.0
info:
  title: BullyStop API
  version: 1.0.0
  description: API para la gestión de usuarios y encuestas en BullyStop
  contact:
    name: DevSociety
    email: support@bullystop.com
servers:
  - url: http://localhost:3000
    description: Servidor de desarrollo
  - url: https://api.bullystop.com
    description: Servidor de producción
components:
  schemas:
    User:
      type: object
      properties:
        id:
          type: integer
          description: ID único del usuario
        name:
          type: string
          description: Nombre completo del usuario
          minLength: 3
        email:
          type: string
          format: email
          description: Correo electrónico del usuario
        grade:
          type: string
          description: Grado escolar del usuario
        emergencyContact:
          type: string
          description: Contacto de emergencia del usuario
          nullable: true
        isActive:
          type: boolean
          description: Estado del usuario
        createdAt:
          type: string
          format: date-time
          description: Fecha de creación del usuario
        updatedAt:
          type: string
          format: date-time
          description: Fecha de última actualización del usuario
      required:
        - name
        - email
        - grade
    CreateUserDto:
      type: object
      properties:
        name:
          type: string
          description: Nombre completo del usuario
          minLength: 3
        email:
          type: string
          format: email
          description: Correo electrónico del usuario
        password:
          type: string
          description: Contraseña del usuario
          minLength: 6
        grade:
          type: string
          description: Grado escolar del usuario
        emergencyContact:
          type: string
          description: Contacto de emergencia del usuario
      required:
        - name
        - email
        - password
        - grade
    UpdateUserDto:
      type: object
      properties:
        name:
          type: string
          description: Nombre completo del usuario
          minLength: 3
        email:
          type: string
          format: email
          description: Correo electrónico del usuario
        password:
          type: string
          description: Nueva contraseña del usuario
          minLength: 6
        grade:
          type: string
          description: Grado escolar del usuario
        emergencyContact:
          type: string
          description: Contacto de emergencia del usuario
...
  responses:
    NotFoundError:
      description: Recurso no encontrado
      content:
        application/json:
          schema:
            type: object
            properties:
              statusCode:
                type: integer
                example: 404
              message:
                type: string
                example: Resource not found
              error:
                type: string
                example: Not Found
    ValidationError:
      description: Error de validación
      content:
        application/json:
          schema:
            type: object
            properties:
              statusCode:
                type: integer
                example: 400
              message:
                type: array
                items:
                  type: string
                example: ["Invalid input data"]
              error:
                type: string
                example: Bad Request
paths:
  /users:
    post:
      summary: Crear un nuevo usuario
      description: Crea un nuevo usuario en el sistema
      tags:
        - Users
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserDto'
      responses:
        '201':
          description: Usuario creado exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          $ref: '#/components/responses/ValidationError'

    get:
      summary: Obtener todos los usuarios
      description: Retorna una lista de todos los usuarios registrados
      tags:
        - Users
      responses:
        '200':
          description: Lista de usuarios
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
  /users/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
        description: ID del usuario

    get:
      summary: Obtener un usuario por ID
      description: Retorna un usuario específico por su ID
      tags:
        - Users
      responses:
        '200':
          description: Usuario encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '404':
          $ref: '#/components/responses/NotFoundError'

    patch:
      summary: Actualizar un usuario
      description: Actualiza los datos de un usuario existente
      tags:
        - Users
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserDto'
      responses:
        '200':
          description: Usuario actualizado exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          $ref: '#/components/responses/ValidationError'
        '404':
          $ref: '#/components/responses/NotFoundError'

    delete:
      summary: Eliminar un usuario
      description: Elimina un usuario del sistema
      tags:
        - Users
      responses:
        '200':
          description: Usuario eliminado exitosamente
        '404':
          $ref: '#/components/responses/NotFoundError'
...